Fixes #3 Reconnect on socket failure
Also fixed up some bugs in the keep alive processing and changed the api so that you can disconnect() and connect() attempt that has not completed.  If the socket experiences an network error, the Listener interface's onDisconnected() method will be called and once connectivity is restored the onConnected() method will be called.